Add files via upload
This project is trying to replicate the behaviour of particles in a complex moving force field. The code was all written by me in the Processing software. In there you will also find a way to convert from hsb colors to rgb. Try tweaking the parameters or improove the performance. There are some essential comments, but if you need help with the code read the readMe and contact me.